requirements.txt
gitpython
PyGithub
PyYAML